+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * [CHANGED] - no longer call exception handler for "ignored" communication exceptions during close/drain after close occurs
 * [ADDED] - #209 - support for comma separated urls in connect() or server()
 * [FIXED] - #206 - incorrect error message when reconnect buffer is overrun
+* [CHANGED] - #214 - moved to an executor instead of 1-off threads
 
 ## Version 2.4.1